feat: a readable arcade HUD with shadowed glyphs and big numerals
Every FIX glyph carries a one pixel drop shadow, which is what keeps the HUD
legible over bright sky and pale dirt alike.  Position, race time, speed and
the countdown use hand-drawn 16x16 numerals kept past tile 255, so the FIX
buffers hold 12-bit tile indices now.  The layout uses the rows the hardware
actually shows - rows 0, 1, 30 and 31 lie outside the picture - with position
and time along the top and speed, gear, stage and course progress in the
bottom corners, clear of the player vehicle.

The start gantry stands sixteen metres ahead of the line, and an object is
only treated as behind a crest with a tolerance that scales with its band.

+++ native/include/game/bajanew.h
@@ -61,8 +61,9 @@ typedef struct BajanewGame {
     int16_t ground_pan;
     /* Aligned so a row can be cleared and compared four cells at a time; the
      * HUD leaves most of every row untouched. */
-    uint8_t fix_shadow[BAJANEW_FIX_ROWS][BAJANEW_FIX_COLUMNS] __attribute__((aligned(4)));
-    uint8_t fix_next[BAJANEW_FIX_ROWS][BAJANEW_FIX_COLUMNS] __attribute__((aligned(4)));
+    /* Cells are 12-bit tile indices: the big numerals live past 255. */
+    uint16_t fix_shadow[BAJANEW_FIX_ROWS][BAJANEW_FIX_COLUMNS] __attribute__((aligned(4)));
+    uint16_t fix_next[BAJANEW_FIX_ROWS][BAJANEW_FIX_COLUMNS] __attribute__((aligned(4)));
     BajaFp progress_scale;
     uint32_t fix_dirty;
     uint32_t fix_written;
